Clean up ISS detector wiring.
