+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/builder/PlatformBuildingBlocks.java
@@ -18,6 +18,7 @@ package com.swirlds.platform.builder;
 
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.common.platform.NodeId;
+import com.swirlds.common.scratchpad.Scratchpad;
 import com.swirlds.platform.consensus.ConsensusSnapshot;
 import com.swirlds.platform.crypto.KeysAndCerts;
 import com.swirlds.platform.event.GossipEvent;
@@ -25,6 +26,7 @@ import com.swirlds.platform.event.preconsensus.PcesFileTracker;
 import com.swirlds.platform.eventhandling.TransactionPool;
 import com.swirlds.platform.gossip.IntakeEventCounter;
 import com.swirlds.platform.recovery.EmergencyRecoveryManager;
+import com.swirlds.platform.state.iss.IssScratchpad;
 import com.swirlds.platform.state.signed.ReservedSignedState;
 import com.swirlds.platform.system.SoftwareVersion;
 import com.swirlds.platform.util.RandomBuilder;
@@ -68,6 +70,7 @@ import java.util.function.Predicate;
  *                                              underlying data source), this indirection can be removed once states are
  *                                              passed within the wiring framework
  * @param initialPcesFiles                      the initial set of PCES files present when the node starts
+ * @param issScratchpad                         scratchpad storage for ISS recovery
  * @param firstPlatform                         if this is the first platform being built (there is static setup that
  *                                              needs to be done, long term plan is to stop using static variables)
  */
@@ -89,4 +92,5 @@ public record PlatformBuildingBlocks(
         @NonNull AtomicReference<Predicate<Instant>> isInFreezePeriodReference,
         @NonNull AtomicReference<Function<String, ReservedSignedState>> latestImmutableStateProviderReference,
         @NonNull PcesFileTracker initialPcesFiles,
+        @NonNull Scratchpad<IssScratchpad> issScratchpad,
         boolean firstPlatform) {}
