+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/builder/PlatformComponentBuilder.java
@@ -19,10 +19,13 @@ package com.swirlds.platform.builder;
 import static com.swirlds.platform.builder.internal.StaticPlatformBuilder.getGlobalMetrics;
 import static com.swirlds.platform.builder.internal.StaticPlatformBuilder.getMetricsProvider;
 import static com.swirlds.platform.gui.internal.BrowserWindowManager.getPlatforms;
+import static com.swirlds.platform.state.iss.IssDetector.DO_NOT_IGNORE_ROUNDS;
 
+import com.swirlds.common.merkle.utility.SerializableLong;
 import com.swirlds.platform.SwirldsPlatform;
 import com.swirlds.platform.components.consensus.ConsensusEngine;
 import com.swirlds.platform.components.consensus.DefaultConsensusEngine;
+import com.swirlds.platform.config.StateConfig;
 import com.swirlds.platform.crypto.CryptoStatic;
 import com.swirlds.platform.crypto.PlatformSigner;
 import com.swirlds.platform.event.creation.DefaultEventCreationManager;
@@ -39,6 +42,7 @@ import com.swirlds.platform.event.orphan.DefaultOrphanBuffer;
 import com.swirlds.platform.event.orphan.OrphanBuffer;
 import com.swirlds.platform.event.preconsensus.DefaultPcesSequencer;
 import com.swirlds.platform.event.preconsensus.DefaultPcesWriter;
+import com.swirlds.platform.event.preconsensus.PcesConfig;
 import com.swirlds.platform.event.preconsensus.PcesFileManager;
 import com.swirlds.platform.event.preconsensus.PcesSequencer;
 import com.swirlds.platform.event.preconsensus.PcesWriter;
@@ -57,6 +61,9 @@ import com.swirlds.platform.event.validation.InternalEventValidator;
 import com.swirlds.platform.eventhandling.DefaultTransactionPrehandler;
 import com.swirlds.platform.eventhandling.TransactionPrehandler;
 import com.swirlds.platform.internal.EventImpl;
+import com.swirlds.platform.state.iss.DefaultIssDetector;
+import com.swirlds.platform.state.iss.IssDetector;
+import com.swirlds.platform.state.iss.IssScratchpad;
 import com.swirlds.platform.state.signed.DefaultSignedStateSentinel;
 import com.swirlds.platform.state.signed.DefaultStateGarbageCollector;
 import com.swirlds.platform.state.signed.ReservedSignedState;
@@ -111,6 +118,7 @@ public class PlatformComponentBuilder {
     private StatusStateMachine statusStateMachine;
     private TransactionPrehandler transactionPrehandler;
     private PcesWriter pcesWriter;
+    private IssDetector issDetector;
 
     /**
      * False if this builder has not yet been used to build a platform (or platform component builder), true if it has.
@@ -793,5 +801,74 @@ public class PlatformComponentBuilder {
             }
         }
         return pcesWriter;
+    }
+
+    /**
+     * Provide an ISS detector in place of the platform's default ISS detector.
+     *
+     * @param issDetector the ISS detector to use
+     * @return this builder
+     */
+    @NonNull
+    public PlatformComponentBuilder withIssDetector(@NonNull final IssDetector issDetector) {
+        throwIfAlreadyUsed();
+        if (this.issDetector != null) {
+            throw new IllegalStateException("ISS detector has already been set");
+        }
+        this.issDetector = Objects.requireNonNull(issDetector);
+        return this;
+    }
+
+    /**
+     * Build the ISS detector if it has not yet been built. If one has been provided via
+     * {@link #withIssDetector(IssDetector)}, that detector will be used. If this method is called more than once, only
+     * the first call will build the ISS detector. Otherwise, the default detector will be created and returned.
+     *
+     * @return the ISS detector
+     */
+    @NonNull
+    public IssDetector buildIssDetector() {
+        if (issDetector == null) {
+            // Only validate preconsensus signature transactions if we are not recovering from an ISS.
+            // ISS round == null means we haven't observed an ISS yet.
+            // ISS round < current round means there was an ISS prior to the saved state
+            //    that has already been recovered from.
+            // ISS round >= current round means that the ISS happens in the future relative the initial state, meaning
+            //    we may observe ISS-inducing signature transactions in the preconsensus event stream.
+
+            final SerializableLong issRound = blocks.issScratchpad().get(IssScratchpad.LAST_ISS_ROUND);
+
+            final boolean forceIgnorePcesSignatures = blocks.platformContext()
+                    .getConfiguration()
+                    .getConfigData(PcesConfig.class)
+                    .forceIgnorePcesSignatures();
+
+            final long initialStateRound = blocks.initialState().get().getRound();
+
+            final boolean ignorePreconsensusSignatures;
+            if (forceIgnorePcesSignatures) {
+                // this is used FOR TESTING ONLY
+                ignorePreconsensusSignatures = true;
+            } else {
+                ignorePreconsensusSignatures = issRound != null && issRound.getValue() >= initialStateRound;
+            }
+
+            // A round that we will completely skip ISS detection for. Needed for tests that do janky state modification
+            // without a software upgrade (in production this feature should not be used).
+            final long roundToIgnore = blocks.platformContext()
+                            .getConfiguration()
+                            .getConfigData(StateConfig.class)
+                            .validateInitialState()
+                    ? DO_NOT_IGNORE_ROUNDS
+                    : initialStateRound;
+
+            issDetector = new DefaultIssDetector(
+                    blocks.platformContext(),
+                    blocks.initialState().get().getState().getPlatformState().getAddressBook(),
+                    blocks.appVersion(),
+                    ignorePreconsensusSignatures,
+                    roundToIgnore);
+        }
+        return issDetector;
     }
 }
