+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/wiring/PlatformSchedulersConfig.java
@@ -59,8 +59,7 @@ import java.time.Duration;
  * @param consensusRoundHandlerSchedulerType       the consensus round handler scheduler type
  * @param consensusRoundHandlerUnhandledCapacity   number of unhandled tasks allowed for the consensus round handler
  * @param runningEventHasher                       configuration for the running event hasher scheduler
- * @param issDetectorSchedulerType                 the ISS detector scheduler type
- * @param issDetectorUnhandledCapacity             number of unhandled tasks allowed for the ISS detector
+ * @param issDetector                              configuration for the ISS detector scheduler
  * @param hashLoggerSchedulerType                  the hash logger scheduler type
  * @param hashLoggerUnhandledTaskCapacity          number of unhandled tasks allowed in the hash logger task scheduler
  * @param completeStateNotifierUnhandledCapacity   number of unhandled tasks allowed for the state completion notifier
@@ -117,8 +116,8 @@ public record PlatformSchedulersConfig(
         @ConfigProperty(defaultValue = "5") int consensusRoundHandlerUnhandledCapacity,
         @ConfigProperty(defaultValue = "SEQUENTIAL CAPACITY(5) UNHANDLED_TASK_METRIC BUSY_FRACTION_METRIC")
                 TaskSchedulerConfiguration runningEventHasher,
-        @ConfigProperty(defaultValue = "SEQUENTIAL") TaskSchedulerType issDetectorSchedulerType,
-        @ConfigProperty(defaultValue = "500") int issDetectorUnhandledCapacity,
+        @ConfigProperty(defaultValue = "SEQUENTIAL CAPACITY(500) UNHANDLED_TASK_METRIC")
+                TaskSchedulerConfiguration issDetector,
         @ConfigProperty(defaultValue = "SEQUENTIAL_THREAD") TaskSchedulerType hashLoggerSchedulerType,
         @ConfigProperty(defaultValue = "100") int hashLoggerUnhandledTaskCapacity,
         @ConfigProperty(defaultValue = "1000") int completeStateNotifierUnhandledCapacity,
