+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/wiring/generate-platform-diagram.sh
@@ -23,8 +23,8 @@ pcli diagram \
     -s 'SelfEventSigner:non-validated events:🍎' \
     -s 'Mystery Input:mystery data:❔' \
     -s 'stateSigner:signature transactions:🖋️' \
-    -s 'issNotificationSplitter:IssNotification:💥' \
-    -s 'toStatusAction:PlatformStatusAction:💀' \
+    -s 'IssDetectorSplitter:IssNotification:💥' \
+    -s 'getStatusAction:PlatformStatusAction:💀' \
     -s 'toNotification:state written notification:📦' \
     -s 'latestCompleteStateNotifier:complete state notification:💢' \
     -s 'OrphanBufferSplitter:preconsensus signatures:🔰' \
@@ -42,7 +42,7 @@ pcli diagram \
     -g 'Preconsensus Event Stream:PcesSequencer,PcesWriter' \
     -g 'Event Creation:EventCreationManager,transactionPool,SelfEventSigner,🍎' \
     -g 'Gossip:gossip,shadowgraph,InOrderLinker' \
-    -g 'ISS Detector:issDetector,issNotificationSplitter,issHandler,toStatusAction' \
+    -g 'ISS Detector:IssDetector,IssDetectorSplitter,issHandler,getStatusAction' \
     -g 'Heartbeat:heartbeat,❤️' \
     -g 'PCES Replay:pcesReplayer,✅' \
     -g 'Consensus Round Handler:consensusRoundHandler,postHandler_stateAndRoundReserver,getState,savedStateController' \
